// src/components/Tabs/enum.ts
/**
 * tab类型
 */
export enum TabType {
  /**
   * 大标题tab
   */
  LargeTitle = 'LargeTitle',

  /**
   * 大文字标签
   */
  LargeText = 'LargeText',

  /**
   * 小文字标签
   */
  SmallText = 'SmallText',
}

/**
 * 排列方式，left 标识自适应宽、有滚动条，center 标识居中、无滚动条
 */
export enum TabAlign {
  left = 'left',
  center = 'center',
}